requirements.txt
# requirements.txt
pyyaml
Faker
